Consolidate rendering of table row - tossed out generic renderer it was a pain to deal with. Having the mouse and draw logic under a single roof is going to be easier and cleaner to deal with.
