+++ crates/core/src/theme/mod.rs
@@ -12,6 +12,7 @@ pub mod picklist;
 pub mod scrollable;
 pub mod text;
 pub mod text_input;
+pub mod table_row;
 
 pub use button::ButtonStyle;
 pub use card::CardStyle;
@@ -21,6 +22,7 @@ pub use modal::ModalStyle;
 pub use picklist::PickListStyle;
 pub use scrollable::ScrollableStyle;
 pub use text_input::TextInputStyle;
+pub use table_row::TableRowStyle;
 
 pub async fn load_user_themes() -> Vec<Theme> {
     log::debug!("loading user themes");
