+++ crates/core/src/widgets/renderer/mod.rs
@@ -1,2 +1,1 @@
 pub mod header;
-pub mod table_row;
